JavaScript with JSX
this is the course selector, true to the table. you can select all of your 1pk-5pk
the only small thing is that you cannot chose math, german, fremdprache into 5pk, even on belibige wahl. so just changing that and all done
some adjustment on the summary so that it can work together with the sql